backend /requirements.txt
fastapi
uvicorn
pymongo
gridfs
python-jose
passlib[bcrypt]
